+++ src/interfaces/envConfig.interface.ts
@@ -1,0 +1,3 @@
+export interface EnvConfigInterface {
+	TELEGRAM_BOT_TOKEN: string;
+}
