--- src/interfaces/envConfig.ts
@@ -1,3 +1,0 @@
-export interface EnvConfig {
-	TELEGRAM_BOT_TOKEN: string;
-}
