+++ src/interfaces/telegram.interface.ts
@@ -1,0 +1,12 @@
+export interface ITelegramChat {
+	id: number;
+}
+
+export interface ITelegramMessageContent {
+	chat: ITelegramChat;
+	text: string;
+}
+
+export interface ITelegramRequest {
+	message: ITelegramMessageContent;
+}
