--- src/interfaces/telegram.ts
@@ -1,12 +1,0 @@
-export interface TelegramChat {
-	id: number;
-}
-
-export interface TelegramMessageContent {
-	chat: TelegramChat;
-	text: string;
-}
-
-export interface TelegramRequest {
-	message: TelegramMessageContent;
-}
